Rename readwrite→_io
