+++ anndata/_core/anndata.py
@@ -1005,7 +1005,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             self._X = None
 
     def _set_backed(self, attr, value):
-        from ..readwrite.utils import write_attribute
+        from .._io.utils import write_attribute
 
         write_attribute(self.file._file, attr, value)
 
@@ -1426,7 +1426,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 dtype=dtype,
             )
         else:
-            from ..readwrite import read_h5ad
+            from .._io import read_h5ad
 
             if filename is None:
                 raise ValueError(
@@ -1923,7 +1923,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             Write sparse data as a dense matrix.
             Defaults to ``True`` if object is backed, otherwise to ``False``.
         """
-        from ..readwrite.write import _write_h5ad
+        from .._io.write import _write_h5ad
 
         if filename is None and not self.isbacked:
             raise ValueError('Provide a filename!')
@@ -1962,7 +1962,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         sep
              Separator for the data.
         """
-        from ..readwrite.write import write_csvs
+        from .._io.write import write_csvs
 
         write_csvs(dirname, self, skip_data=skip_data, sep=sep)
 
@@ -1975,7 +1975,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         filename
             The filename.
         """
-        from ..readwrite.write import write_loom
+        from .._io.write import write_loom
 
         write_loom(filename, self, write_obsm_varm=write_obsm_varm)
 
@@ -1995,7 +1995,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         chunks
             Chunk shape.
         """
-        from ..readwrite.write import write_zarr
+        from .._io.write import write_zarr
 
         write_zarr(store, self, chunks=chunks)
