+++ anndata/tests/test_backed_sparse.py
@@ -55,7 +55,7 @@ def test_dataset_append_memory(tmp_path, sparse_format, append_method):
     b = sparse_format(sparse.random(100, 100))
 
     with h5py.File(h5_path, "a") as f:
-        ad.readwrite.h5ad.write_attribute(f, "mtx", a)
+        ad._io.h5ad.write_attribute(f, "mtx", a)
         diskmtx = SparseDataset(f["mtx"])
 
         diskmtx.append(b)
@@ -79,8 +79,8 @@ def test_dataset_append_disk(tmp_path, sparse_format, append_method):
     b = sparse_format(sparse.random(10, 10))
 
     with h5py.File(h5_path, "a") as f:
-        ad.readwrite.h5ad.write_attribute(f, "a", a)
-        ad.readwrite.h5ad.write_attribute(f, "b", b)
+        ad._io.h5ad.write_attribute(f, "a", a)
+        ad._io.h5ad.write_attribute(f, "b", b)
         a_disk = SparseDataset(f["a"])
         b_disk = SparseDataset(f["b"])
 
@@ -105,8 +105,8 @@ def test_wrong_shape(tmp_path, sparse_format, a_shape, b_shape):
     b_mem = sparse.random(*b_shape, format=sparse_format)
 
     with h5py.File(h5_path, "a") as f:
-        ad.readwrite.h5ad.write_attribute(f, "a", a_mem)
-        ad.readwrite.h5ad.write_attribute(f, "b", b_mem)
+        ad._io.h5ad.write_attribute(f, "a", a_mem)
+        ad._io.h5ad.write_attribute(f, "b", b_mem)
         a_disk = SparseDataset(f["a"])
         b_disk = SparseDataset(f["b"])
 
@@ -119,7 +119,7 @@ def test_wrong_formats(tmp_path):
     base = sparse.random(100, 100, format="csr")
 
     with h5py.File(h5_path, "a") as f:
-        ad.readwrite.h5ad.write_attribute(f, "base", base)
+        ad._io.h5ad.write_attribute(f, "base", base)
         disk_mtx = SparseDataset(f["base"])
         pre_checks = disk_mtx.to_memory()
