+++ anndata/tests/test_utils.py
@@ -2,7 +2,7 @@ import pytest
 import zarr
 import h5py
 
-from anndata.readwrite.utils import report_read_key_on_error, AnnDataReadError
+from anndata._io.utils import report_read_key_on_error, AnnDataReadError
 
 
 def test_key_error(tmp_path):
